+++ cs3560cli/commands/create_gitignore.py
@@ -6,27 +6,30 @@ from https://github.com/github/gitignore, it will also
 add all OS dependent .gitignore as well.
 """
 
+import itertools
 from pathlib import Path
 
 import click
 import requests
 
-# TODO: add more mappings.
-path_mapping = {
+ALIASES = {
     "windows": "Global/Windows.gitignore",
     "macos": "Global/macOS.gitignore",
     "vscode": "Global/VisualStudioCode.gitignore",
     "python": "Python.gitignore",
+    "notebook": "community/Python/JupyterNotebooks.gitignore",
     "cpp": "C++.gitignore",
+    "c++": "C++.gitignore",
+    "c": "C.gitignore",
     "node": "Node.gitignore",
-    "unity": "Unity.gitignore",
+    "js": "Node.gitignore",
+    "java": "Java.gitignore",
+    "kotlin": "Java.gitignore",
+    "go": "Go.gitignore",
     "rust": "Rust.gitignore",
-}
-
-aliases = {
-    "c++": "cpp",
-    "js": "node",
-    "VisualStudioCode": "vscode",
+    "unity": "Unity.gitignore",
+    "tex": "TeX.gitignore",
+    "latex": "TeX.gitignore",
 }
 
 
@@ -34,102 +37,136 @@ class ApiError(Exception):
     pass
 
 
-def get_path(name: str) -> str | None:
-    """Lookup the path.
-
-    Assume name is in lowercase and is not an alias.
-    """
-    if name not in path_mapping.keys():
-        return None
-
-    return path_mapping[name]
-
-
-def normalize(name: str) -> str:
-    """Apply alias and lower the case the name."""
-    name = name.lower()
-
-    if name in aliases.keys():
-        name = aliases[name]
-
-    return name
-
-
-def get_content(
+def build_gitignore_content(
     names: list[str],
     bases: list[str] | None = None,
     root: str = "https://raw.githubusercontent.com/github/gitignore/main/",
     header_text_template: str = "#\n# {path}\n# Get the latest version at https://github.com/github/gitignore/{path}\n#\n",
-) -> str:
-    """Create .gitignore content from list of names.
-
-    Assume that names in bases area already normalized.
-    """
+) -> tuple[str, bool]:
+    """Create .gitignore content from list of names and bases."""
     if bases is None:
         bases = ["windows", "macos"]
+    else:
+        bases = [name.lower() for name in bases]
 
     final_text = ""
+    names = bases + [name for name in names if name.lower() not in bases]
 
-    names = bases + [normalize(name) for name in names if normalize(name) not in bases]
-
+    error_occured = False
     for name in names:
         if name is None:
             continue
-        path = get_path(name)
-        if path is None:
-            continue
+        path = ALIASES.get(name.lower(), name)
+        url = root + path
+
         try:
-            res = requests.get(root + path)
+            click.echo(f"Fetching '{name}' from {url} ...")
+            res = requests.get(url)
             if res.status_code == 200:
                 header_text = header_text_template.format(path=path)
                 final_text += header_text
                 final_text += res.text + "\n"
             else:
-                raise ApiError(
-                    f"status code from API is not as expected. Expect 200 but get {res.status_code}"
+                click.echo(
+                    f"[error]: failed to fetch '{name}' (HTTP code: {res.status_code}). It will be skipped."
                 )
+                error_occured = True
         except requests.exceptions.RequestException as e:
             raise ApiError("error occur when fetching content") from e
-    return final_text
+    return final_text, error_occured
 
 
 @click.command("create-gitignore")
 @click.argument("names", type=str, nargs=-1)
-@click.option("--root", type=click.Path(exists=True), default=".")
-@click.option("--base", type=str, multiple=True, default=("windows", "macos"))
+@click.option(
+    "--outfile",
+    "-o",
+    type=click.Path(exists=False, dir_okay=False),
+    default=".gitignore",
+    help="Specify an output file.",
+)
+@click.option(
+    "--base",
+    "-b",
+    type=str,
+    multiple=True,
+    default=("windows", "macos"),
+    help='Base content of the file. Default: windows, macos. To not include any base content, use --base "". To specify multiple base, use --base name1 --base name2 ...',
+)
+@click.option(
+    "--list-mapping",
+    "-l",
+    type=bool,
+    is_flag=True,
+    help="Show list of available mappings.",
+)
 @click.pass_context
 def create_gitignore(
     ctx: click.Context,
     names: list[str],
-    root: str | Path = ".",
+    outfile: str | Path = ".",
     base: list[str] | tuple[str, ...] = ("windows", "macos"),
+    list_mapping: bool = False,
 ) -> None:
-    """Create .gitignore content from list of names.
+    """Create .gitignore content from list of NAMES.
 
-    Assume that names in bases area already normalized.
+    The windows and macos content for .gitignore will be added by default. Use --base "" to disable this. The command will fetch the content
+    from github/gitignore repository on GitHub using https://raw.githubusercontent.com/github/gitignore/main/.
+
+    --list-mapping can be used to view avaialble mapping. If the provided NAMES is not part of the mappings, it will be used as is.
     """
-    if isinstance(root, str):
-        root = Path(root)
+    if list_mapping:
+        click.echo("The following aliases are available:")
+        for key in ALIASES:
+            click.echo(f"- {key} -> {ALIASES[key]}")
+        ctx.exit()
+
+    if isinstance(outfile, str):
+        outfile = Path(outfile)
     if isinstance(names, tuple):
         names = list(names)
     if isinstance(base, tuple):
-        base = list(base)
+        bases = list(base)
+
+    bases = [name for name in bases if len(name.strip()) != 0]
+
+    name_text = "\n".join(
+        itertools.chain(
+            [f"- {name} (from bases)" for name in bases],
+            [f"- {name}" for name in names],
+        )
+    )
+
+    if len(names) == 0 and len(bases) == 0:
+        click.echo(f"Will create an empty '{outfile.name}' file at {outfile.parent}")
+    else:
+        click.echo(
+            f"Will create '{outfile.name}' file at {outfile.parent} with the following content:\n{name_text}"
+        )
+    click.confirm("Do you want to continue?", default=True, abort=True)
+
     try:
-        content = get_content(names, base)
+        content, error_occured = build_gitignore_content(names, bases)
+        if error_occured:
+            click.confirm(
+                "One or more name failed to fetch. Do you want to continue?",
+                default=False,
+                abort=True,
+            )
     except ConnectionError as e:
         ctx.fail(f"network error occured\n{e}")
     except ApiError as e:
         ctx.fail(f"api error occured\n{e}")
 
-    outfile_path = root / ".gitignore"
-    if outfile_path.exists():
-        ans = click.confirm(f"'{outfile_path!s}' already exist, overwrite?")
+    if outfile.exists():
+        ans = click.confirm(f"'{outfile!s}' already exist, overwrite?")
         if not ans:
-            click.echo(f"'{outfile_path!s}' is not modified")
+            click.echo(f"'{outfile!s}' is not modified")
             ctx.exit()
 
-    with open(outfile_path, "w") as out_f:
+    with open(outfile, "w") as out_f:
         out_f.write(content)
+    click.echo(f"Content is written to '{outfile!s}'")
 
 
 if __name__ == "__main__":